7 failed basic tests 42FileChecker

+++ libft/ft_count.c
@@ -6,15 +6,15 @@
 /*   By: aschukin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/27 19:31:15 by aschukin          #+#    #+#             */
-/*   Updated: 2018/03/06 15:56:29 by aschukin         ###   ########.fr       */
+/*   Updated: 2018/03/27 15:35:19 by aschukin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_count(long long num)
+intmax_t	ft_count(long long num)
 {
-	size_t count;
+	intmax_t	count;
 
 	count = 0;
 	if (num <= 0)
